refactor: consolidate player types
