+++ src/Data/aflPlayers.ts
@@ -1,4 +1,4 @@
-import type { Player } from '../types';
+import type { Player } from '../types/players';
 
 const aflPlayers: Player[] = [
   {
