+++ src/components/DraftOrderBar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { Team } from '../types';
+import type { Team } from '../types/players';
 
 interface DraftOrderBarProps {
   teams?: Team[];
