+++ src/components/PlayerDetail.tsx
@@ -1,4 +1,4 @@
-import type { Player } from '@/types';
+import type { Player } from '@/types/players';
 import PlayerStatsDisplay from './PlayerStatsDisplay';
 
 interface PlayerDetailProps {
