+++ src/components/PlayerFilters.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useMemo } from 'react';
-import type { Player } from '../types';
+import type { Player } from '../types/players';
 
 type PlayerFiltersProps = {
   players: Player[];
