+++ src/components/PlayerSummaryCard.tsx
@@ -2,7 +2,7 @@
 'use client';
 
 import React from 'react';
-import type { Player } from '../types';
+import type { Player } from '../types/players';
 
 type Props = {
   player: Player;
