+++ src/components/TradeCentreClient.tsx
@@ -3,7 +3,7 @@
 import { useMemo, useState, useCallback, useEffect } from 'react';
 import Link from 'next/link';
 import { useDebounce } from '@/Hooks/useDebounce';
-import type { Player } from '@/types';
+import type { Player } from '@/types/players';
 import { statLabels, TradeCentreStrings } from '@/lib/constants';
 import { useTradeStore } from '@/state/tradeStore';
