+++ src/components/TradeCentreShell.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useMemo, useState } from 'react';
-import type { Player } from '@/types';
+import type { Player } from '@/types/players';
 import OfferDock from '@/components/OfferDock';
 import { Column as TeamColumn } from '@/components/SideBySideTeams';
