+++ src/lib/data.ts
@@ -2,7 +2,7 @@
 import 'server-only';
 import { promises as fs } from 'node:fs';
 import path from 'node:path';
-import type { Player } from '@/types';
+import type { Player } from '@/types/players';
 
 type AnyObj = Record<string, unknown>;
