+++ src/scrapeFootywireStats.ts
@@ -3,7 +3,7 @@ import * as cheerio from 'cheerio';
 import type { Element } from 'domhandler';
 import { db } from '@/lib/firebaseClient';
 import { collection, addDoc } from 'firebase/firestore';
-import type { Player } from './types';
+import type { Player } from './types/players';
 
 // Replace with any valid match ID
 const matchId = '11341'; // Example: A recent match
